Move skip method from e2e fw ginkgowrapper to e2e skipper fw
